IPython interpreter
 Python-bash, script and show ( functions for each point):
At this point, you are now a programmer capable of writing simple functions in both Bash and Python. Programmers learn by doing, though, so at this point we highly recommend that you change the system calls in these two Bash and Python programs and make them your own. Give yourself some bonus points if you add several new functions to the script and call them from a main function.

Open the IPython and Bash windows if you closed them so we can quickly demonstrate why the functions are important for code reuse. One of the first scripts we created in Python was a script in a file called pysysinfo.py. In Python, since a file is a module and vice versa, we can import this script file into IPython. Note that you never need to specify the .py part of the file you are importing. In fact, if you do this, the import will not work. This is what the bash terminal or code looks like.

The ports that are ‘listening’ in your system,Disk usage in your system,Network interfaces
and Routing table.